requirements.txt
accelerate==0.16.0
datasets==2.9.0
diffusers==0.16.1
evaluate==0.4.0
huggingface-hub==0.13.4
importlib-resources==5.10.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
mit-semseg==1.0.0
-e git+https://github.com/IntelLabs/FP8-Emulation-Toolkit.git@57177ca90cd7947762f21016ca9986625da67f7e#egg=mpemu
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
pytorch-quantization==2.1.2
qtorch==0.3.0
scikit-learn==1.2.1
scipy==1.10.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tokenizers==0.13.2
torch==1.13.0+cu116
torchaudio==0.13.1
torchmetrics==0.11.4
torchview==0.2.6
torchvision==0.14.0+cu116
tqdm==4.64.1
transformers==4.27.4
